+++ src/level/enemy/BorissaurEnemy.java
@@ -15,7 +15,7 @@ public class BorissaurEnemy extends Enemy {
      * A dumb hostile enemy native to Mars
      */
     @Reflective
-    public BorissaurEnemy(Map map, RealBasebot bot, int x, int y) {
+    public BorissaurEnemy(Map map, RealBasebot bot, float x, float y) {
         super(map, bot, CIRCLE, x, y, true);
         width = TILE_SIZE * 0.7f;
         height = TILE_SIZE * 0.7f;
